# api/openapi/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/me/inbox/messages/comments:
        get:
            tags:
                - MessageService
            description: GET /api/v1/me/inbox/messages/comments 当前用户评论消息列表
            operationId: MessageService_ListCommentInboxMessages
            parameters:
                - name: readFilter
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.ListCommentInboxMessagesResponse'
    /api/v1/me/inbox/messages/follows:
        get:
            tags:
                - MessageService
            description: GET /api/v1/me/inbox/messages/follows 当前用户关注消息列表
            operationId: MessageService_ListFollowInboxMessages
            parameters:
                - name: readFilter
                  in: query
                  schema:
                    type: integer
                    format: enum
                - name: pageToken
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.ListFollowInboxMessagesResponse'
...
    /api/v1/me/inbox/messages/{uid}:
        delete:
            tags:
                - MessageService
            description: DELETE /api/v1/me/inbox/messages/{uid} 归档一条消息
            operationId: MessageService_DeleteInboxMessage
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/message.DeleteInboxMessageResponse'
...
components:
    schemas:
...
        message.CommentInboxMessage:
            required:
                - uid
                - isRead
                - actor
                - createdAt
            type: object
            properties:
                uid:
                    type: string
                isRead:
                    type: boolean
                actor:
                    $ref: '#/components/schemas/message.InboxMessageActor'
                createdAt:
                    type: string
                commentUid:
                    type: string
                commentContent:
                    type: string
                postUid:
                    type: string
                parentUid:
                    type: string
                parentContent:
                    type: string
...
        message.DeleteInboxMessageResponse:
            type: object
            properties: {}
        message.FollowInboxMessage:
            required:
                - uid
                - isRead
                - actor
                - createdAt
            type: object
            properties:
                uid:
                    type: string
                isRead:
                    type: boolean
                actor:
                    $ref: '#/components/schemas/message.InboxMessageActor'
                createdAt:
                    type: string
        message.InboxMessageActor:
            required:
                - uid
                - nickname
                - avatarUrl
            type: object
            properties:
                uid:
                    type: string
                nickname:
                    type: string
                avatarUrl:
                    type: string
        message.ListCommentInboxMessagesResponse:
            required:
                - messages
                - nextPageToken
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/message.CommentInboxMessage'
                nextPageToken:
                    type: string
        message.ListFollowInboxMessagesResponse:
            required:
                - messages
                - nextPageToken
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/message.FollowInboxMessage'
                nextPageToken:
                    type: string
...